pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "literature-ingestor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # PDF parsing
    "marker-pdf>=1.6",
    # Vector store
    "chromadb>=0.6",
    # LM Studio / OpenAI-compatible client
    "openai>=1.57",
    # Metadata store
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    # Config
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    # Utilities
    "tqdm>=4.67",
    "rich>=13.9",
    "typer>=0.15",
]

[project.scripts]
lit = "literature_ingestor.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
